chore(StatusDot): rename folder
